fix: address review findings on #1259 — move chat-poller timer tests to integration layer
The start/stop timer-lifecycle tests used real ~100ms setTimeout waits,
putting them over the unit-layer <200ms hard cap and outside its no-I/O
boundary (docs/test-readiness/test-system.md). Bun's test runner has no
Jest-style fake-timer API, so the tests are moved as-is to a new
chat-poller.integration.test.ts rather than faked.
